+++ tests/testthat/test-provider-google.R
@@ -76,7 +76,7 @@ test_that("can use pdfs", {
 test_that("can match prices for some common models", {
   provider <- chat_google_gemini_test()$get_provider()
 
-  expect_true(has_cost(provider, "gemini-2.5-flash"))
+  expect_true(has_cost(provider, "gemini-3.5-flash"))
 
   expect_false(has_cost(provider, "gemini-1.0-pro-latest"))
 })